Compatability fixes from Sail 0.7.1 to Sail 0.8
Two small compatability fixes:

Small change to the prelude to make sure we are forwards compatable
with the next release of Sail, which has slightly different syntax for
implicit arguments.

Due to changes in the monad embedding in the latest git version of
Sail, we disable generating Lem by default when we detect we are on
that version. It's a small one line fix to correct, but we want to
keep the Lem definitions in this repository compatible with the opam
release 0.7.1 for now.