Work around Isabelle problem in PTW functions
The standard pattern completeness proof for recursive functions
generated by Lem for Isabelle seems to get confused in some situations
when there are variables with unit type (in particular the ext_ptw
argument of the page table walk functions).  Solving this properly
requires some more digging into the Isabelle simplifier, but the ad-hoc
workaround in this commit fixes the problem for now.