Fix parsing long options in the C emulator for RVFI-DII
idx, passed as longindex to getopt_long, is updated every time a long
option is parsed to be the index of that option in the options array,
not argv. We should instead use optind as in the non-RVFI-DII case, and
we can remove the variable to avoid confusion since it is unused.