Allow extensions to override page fault causes
Otherwise there's no use for PTW_Ext_Error. This is required for
sail-cheri-riscv to be able to make use of its own page table
exceptions. The alternative to this is to pass the ext_ptw to
translationException, but given PTW_Ext_Error exists it seems this was
the intended way, and aligns with how access faults vs page faults are
distinguished.