A kinder, gentler splitting of ext_access_type's ReadWrite
This redoes https://github.com/rems-project/sail-riscv/pull/57 without
nearly as much excitement.  We sill want it for CHERI, so that we can
signal from the instruction to the PTW whether we are prepared to load a
capability (or will strip any tags that we load) and whether the store
will (or might) set a tag.

Thanks to Prashanth Mundkur for several improvements.