vmem_rvNN: allow shimming underneath effectivePrivilege
It's possible that we might want to do page table walks at Privilege levels
other than the default value (e.g., under the explicit direction of the
instruction stream).  Split translateAddr into a thin wrapper of the same name
and a new translateAddr_priv that takes the Privilege as an argument.