Support D extension on RV32 (#108)
* Use bool for floating point comparison result

Using bits_WU (bits(32)) or bits_LU (bits(64)) makes no sense, these are
just boolean values, and having fixed-width types is a pain for
suporting RV32D (since RV32D would need to truncate, but RV128D would
need to extend). Instead represent these as an actual bool to match what
the values really are. This could be done with bits(1) but the value is
logically a boolean (like the built-in integer comparison operators) not
a bit vector of length one so we convert to bool and back for a cleaner
interface.

* Support compiling RV32F with flen == 64

The code conflated flen and xlen; deconflating them and adding suitable
assertions (rather than silently retiring as a no-op that doesn't bump
instret) ensures that it can be compiled for RV32 with flen set to 64.
Whilst here, add the extensions and truncations that would be needed for
RV128F.

Note that there are already suitable guards on the decode clauses to
ensure these instructions are illegal on RV32.

* Support compiling RV32D

This copies various bits of XLEN generality from the F code.

* Support RV32D loads/stores

* Correctly initialise misa.D based on flen not xlen

* Makefile: Enable D extension for RV32

This now works so can be enabled.

* test: Enable RV32D tests