Fix handling of mstatus.MIE in M-mode only case and remove incorrect assert.
Previously if neither user or supervisor mode are present
dispatchInterrupt ignored mstatus.MIE leading to an infinite interrupt
loop. dispatchInterrupt handles the case where no delegated interrupts
are possible (e.g. M-mode only) separately and I think this branch was
untested. It contained an incorrect assert that was falsifiable on an
CPU with M and U modes but no S-mode or N-extension. I've removed this
assert and simplified the check for possible interrupt delegation.