riscv_sim: Add command line arguments to enable or disable use of boot ROM.
This adds command line arguments --boot-rom and --no-boot-rom to
enable or disable the use of the boot ROM at the reset vector.  The
current behaviour (--boot-rom) initialises a small bit of memory at
the reset vector to set up argument registers and then jump to the
address specified by entry point of the ELF file.  With --no-boot-rom
the simulator just sets the initial PC to the entry point.  For
backwards compatibility the default is to use the boot-rom unless
NO_BOOT_ROM is defined during compilation.