zvkned: add vaeskf1.vi
The "vaeskf1.vi" instruction performs a single round
of the forward AES-128 KeySchedule generation.

The next round key is generated word by word from the current round
key element group in "vs2" and the immediately previous word of the
round key. The least significant word is generated using the most
significant word of the current round key as well as a round constant
which is selected by the round number.

The round number(ranges 1-10), comes from imm[3:0]; imm[4] is ignored.
We project out-of-range immediates onto in-range values by inverting
imm[3].

Signed-off-by: Charalampos Mitrodimas <charalampos.mitrodimas@vrull.eu>