Zvksh: add "vsm3c.vi" instruction
Vector SM3 Compression, Two rounds of SM3 compression are performed.

The current state of eight 32-bit words is read in as an element group
from vd. Eight 32-bit message words are read in as an element group
from vs2, although only four of them are used. All of the 32- bit
input words are byte-swapped from big endian to little endian. These
inputs are processed somewhat differently based on the round group (as
specified in rnds), and the next state is generated as an element
group of eight 32-bit words. The next state of eight 32-bit words are
generated, swapped from little endian to big endian, and are returned
in an eight-element group.

Signed-off-by: Charalampos Mitrodimas <charalampos.mitrodimas@vrull.eu>