Zfa: rewrite fcvtmod.w.d in functional style
The original implementation of fcvtmod.w.d (in fcvtmod_helper) used
both an imperative style and suffered from the following shortcomings:
- it required extra work (i.e., it first converted to a 64bit value)
- the conversion from a normal fixed-point significand to the 64bit
  value (and the raising of exception flags) was hard to follow (and
  hard to verify)

This rewrites the helper function in a purely functional style and
performs the conversion from a fixed-point (53 bit precision) number
to a 32 bit integer using a single exponentiation (i.e. a single shift
operation).

Signed-off-by: Philipp Tomsich <philipp.tomsich@vrull.eu>
Signed-off-by: Charalampos Mitrodimas <charalampos.mitrodimas@vrull.eu>