Update pre-commit-hooks version
The template that I used when adding it uses hooks from 2020. Update this
to the latest version and add a minimum pre-commit version check to ensure
we get sensible error messages rather than obscure failures.
The minimum chosen here is 2.10 as that is the version shipping with
Debian 11 and most other distributions have newer versions. If needed
a newer version can always be installed using pip.