Ssdtso: Add DTSO flag to {m,s,h}envcfg
This patch introduces the DTSO field to the {m,s,h}envcfg
register as defined in the Ssdtso specification.

During the ratification process of Ztso it was agreed that
Sail does not support TSO and will not be support it going
forward.

Therefore, this patch does not use the DTSO field anywhere.

Signed-off-by: Christoph Müllner <christoph.muellner@vrull.eu>