Simplify boxed float register access functions
This replaces the separate functions to read/write the floating point registers as smaller widths (which boxes/unboxes them) with single generic functions.

Note in some cases we still need to explicitly pass the width in so it might seem like there is no benefit, but those will be removed once the code is fully generic.