Add Zfhmin support
Per section 24.6 of the unprivileged ISA manual:

> The Zfhmin extension includes the following instructions from the Zfh extension: FLH, FSH, FMV.X.H, FMV.H.X, FCVT.S.H, and FCVT.H.S. If the D extension is present, the FCVT.D.H and FCVT.H.D instructions are also included. If the Q extension is present, the FCVT.Q.H and FCVT.H.Q instructions are additionally included.

We already supported Zfh, so this is just a modification of some of the guards to enable only the supported instructions for Zfhmin. For now I have Zfhmin enabled whenever Zfh is enabled, but later it can be updated so that it can be enabled independently once we have better model configuration.