Zero mstatus.UPIE/UIE bits in legalize_mstatus
This was missed when we removed the N extension.

I'm not really a fan of this code style - in other places we explicitly copy the bits that *are* implemented using their field names, which is also not ideal but definitely better and less error prone. We should change this code to do that at some point. I added a TODO.