Fix delegated interrupt handling
When the N extension was removed this code was accidentally broken because it doesn't respect mie for delegated interrupts any more.

The `internal_error`s should be unreachable, at least for the sequential output. I believe formal backends may decide `mideleg` is going to be non-zero even without S, but I think that was an existing issue.