Make vector length (vl) calculation configurable
When `vl` is recalculated due to `vset[i]vl[i]`, the implementation can have some freedom about which value to use. Specifically when VLMAX < AVL < 2*VLMAX, the implementation can choose any `vl` such that ceil(AVL / 2) <= vl <= VLMAX. See "Constraints on Setting vl" in the spec for more details.

This adds a new option that allows you to pick between `ceil(AVL / 2)` and `VLMAX`. It is not yet hooked up to a CLI option but we will do that once the new config system is in place.

Note that this changes the default from `ceil(AVL / 2)` to `VLMAX`.

Co-authored-by: Yui5427 <785369607@qq.com>