Support static compilation of the emulator
Optionally wrap the libs with `-Wl,-Bstatic`, `-Wl,-Bdynamic` which will enable static compilation for those libs. We don't really want to use `-static` because it statically links glibc, which is a bad idea, and also increases the binary size. Linking with a sufficiently old version of glibc is a better option.

On Mac it dynamically links with zlib, but that's acceptable since zlib is always provided by MacOS. Also Zlib has been removed as a dependency for the next Sail compiler release.