Make physical addresses 34-bit for RV32
Change physical addresses to be 34-bits on RV32. Sv32 allows 34-bit physical addresses. When no virtual memory is supported or the Bare translation mode is used, the 34-bit addresses are just zero extended from the 32-bit virtual address.

>  When the value of satp.MODE is Bare, the 32-bit virtual address is translated (unmodified) into a 32-bit physical address, and then that physical address is zero-extended into a 34-bit machine-level physical address.

The underlying Sail memory system only supports 32 or 64-bit addresses, so now we zero-extend to 64-bit in all cases.

Note, Sv32 can be used on RV64 if SXLEN=32. The model does not support SXLEN != XLEN yet but this code may need modifying if support is added.