Remove notes_FD_extensions
These are the original notes from the implementation of the F and D extensions. They've fallen fairly out of date already and will become even more obsolete as we switch to the pure Sail float implementation. This documentation seems to be of limited value at this point, so rather than updating it I'm just removing it.