test: do not clean between tests
There is a bug in the build process if we need to make clean between
builds.  We originally needed make clean to avoid running test with the
old binary, but that has been fixed.

Run the tests without wasting time by rebuilding the binaries.

The tests now run much faster, so maybe we can finally require
```git rebase -x './tests/run_tests.sh' origin/master```
before submitting patches and do it for pull requests.
(Testing for regressions is slow even after this if the patch series is
 also supposed to be properly split.)

Signed-off-by: Radim Krčmář <rkrcmar@ventanamicro.com>