Improve Zicbom implementation.
Clarify the privilege checking of xenvcfg.CBIE in cbo.inval.
The privilege check should extend to the hypervisor case, when
two possible illegal exceptions could be thrown.  This check as
written doesn't match the pseudo-Sail in the CMO spec, but is
arguably clearer.

Make sure that the reserved CBIE value is not representable.