Refactor instruction retirement.
The eventual goal here is to surface wait-states of the hart
(e.g. as caused by WFI and WRS.{NTO,STO}) into the stepper,
and handle debug triggers from execution for extensions such as
Sdtrig.

The refactor makes interrupt and exception handling more
symmetrical: both are now handled in the stepper.

This also removes the model-internal global instbits register
into a stepper-local variable, which fixes #412.

This refactor is based on ideas discussed in #398 and #412.

Co-authored-by: Tim Hutt <timothy.hutt@codasip.com>
Co-authored-by: Jessica Clarke <jrtc27@jrtc27.com>
Co-authored-by: Alasdair Armstrong <alasdair.armstrong@cl.cam.ac.uk>