Implement a Wait state in the Sail stepper.
The Wait state is exited on interrupts or if requested by the C
emulator.  The current implementation requests an immediate exit.

This adds an option to treat the Wait as a NOP.  This is currently
only used for WFI, but it could be re-used for WRS.{STO, NTO} in
the future.