Add first party C tests using Zig for cross-compilation
Add first party C tests that can be used for tests specific to this model. This uses Zig as the compiler since it's small and easy to set up. We could automatically download it for users like we can for GMP (I haven't implemented that yet).

This includes a copy of `nanoprinf` for printing to the UART. It has two tests, `test_hello_world.c`, which passes, and `test_max_pmp.c` which fails due to a model bug.

There are some TODOs and a lot of it needs cleaning up a fair but bit I wanted some feedback before I continue.