Store `total_insns` and `insn_limit` as `uint64_t`, not `int`
At a few hundred thousand instructions per second, 32-bit `int total_insns` will overflow in a few hours. If the limit is too high, it will be parsed incorrectly without error and execution stops at a somewhat arbitrary point (or not at all).

This commit therefore changes the types to `long`. 

I've also switched to using `strtol` for parsing the `insns_limit` command line argument in order to get error checking (and to make it possible to use hex notation).