Implement Zimop and Zcmop extensions
One of the tricky parts with this extension is that it is designed to have its instructions overridden by other extensions (like `Zicfiss`). This will be handled by relying on file order to override instructions.

This is a little fragile with the current compilation method, where all `.sail` files are passed to the compiler. In future we will use a Sail project file which allows enforcing file order more robustly.