Add cmake support for Rocq
This includes changing the prover name from Coq to its new name Rocq, updating the opam package file, adding a couple of built-in functions to the handwritten portion, and building the resulting Rocq code. The output is all gathered in one rocq build directory to keep the namespace manageable and convenient.

It also drops the extra `_variant` from the name of lem/rmem/rocq output, which is rather redundant.