Replace ad-hoc model configuration with Sail-native configuration
This replaces the ad-hoc C stubs used for configuration with a Sail-native configuration system using a separate configuration file. With the native configuration system adding a new configurable option to the model is easy, add `config x.y.z : T` anywhere in the model, and then add `x.y.z` into the JSON file, and it'll just work.

This removes a lot of CLI options to configure the model, and moves them into a single `-c/--config config.json` option. A default configuration (`config/default.json`) is used if that option isn't set. `-c` was previously used for coverage but it wasn't compiled in by default.

This updates the model to use the new config system, but doesn't add any additional configuration settings. The only behaviour change is that the handling of `rom_base` and `rom_size` is slightly different, in that it'll just print if the configuration values don't match what the C code that creates the reset vector and writes the DTB expect rather than setting the values.

Some tasks still need to be done:

1. Update the other Sail targets. This configuration system works across all Sail backends, but the CMake commands for them need to be updated to reflect this.
2. Sail can generate a JSON schema that will validate any provided config fits into the model, that hasn't been set up yet. The default config should be JSON schema validated in CI as well.
3. A friendly way to generate the JSON, e.g. a script to import riscv-config/UDB etc.